Add the DataLand runtime system
This refactor introduces the DataLand runtime system that
encapsulates the interpreter, the data, and the visualization
spec. The React app accesses the runtime state and methods
though a Higher-Order Component. This change also introduces
a DataTable object type that is responsible for all the data
transformation that the project code might request.