Initialize Blockly through initBlockly()
This change decouples the initialization of Blockly (i.e., custom
block definitions and other customizations) to outside of the
React component. Apps using this component will have to call initBlockly()
separately before using the React component. This approach avoids
multiple initializations of Blockly if multiple instances of the Editor
component is being used in a single page.